\name{MonoInc-package}
\alias{MonoInc-package}

\docType{package}
\title{
Cleans and imputes on monotonic data.

}
\description{
The MonoInc package in R seeks to clean data so that erroneous values are less effective statistically.  Given a prespecified range, MonoInc will determine if an observation is “unusual”, and then replace the value at the user's will.  MonoInc will impute on participant data individually, so that the number of time points need not be the same. MonoInc will also remove duplicate rows.
}
\details{
\tabular{ll}{
Package: \tab MonoInc\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2016-05-19\cr
License: \tab GPL-3\cr
}

}
\author{Melyssa Minto, Michele Josey


Maintainer: Michele Josey \href{mailto:mjosey@nccu.edu}{mjosey@nccu.edu} 
}

\keyword{ package }
\keyword{ MonoInc }
\keyword{mono }
\keyword{monotonic }
